Make it compatible with as back as 5.5
